% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_missing_data}
\alias{calculate_missing_data}
\title{calculate_missing_data}
\usage{
calculate_missing_data(df_list, group, subset)
}
\arguments{
\item{df_list}{[[list]] list of data.frames}

\item{group}{[[character]] group variable name \itemize{
\item option1
\item option2
\item option3}}

\item{subset}{[[character]] subset name, currently only `rand source`}
}
\value{
[[data.frame]]  calculated data.frame
}
\description{
Function calculates missing data number and ratio in groups after subsetting
}
