% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_number_and_ratio}
\alias{calculate_number_and_ratio}
\title{calculate_number_and_ratio}
\usage{
calculate_number_and_ratio(df, group = NULL, var)
}
\arguments{
\item{df}{[[data.frame]] data}

\item{group}{[[character]] group variable name,
\itemize{
\item option1
\item option2
\item option3},
 by default `NULL`}

\item{var}{[[character]] variable to summarise}
}
\value{
[[data.frame]]  summarised data.frame
}
\description{
Function calculates number and ratio of given variable which can be grouped
}
