% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_rand_source}
\alias{calculate_rand_source}
\title{calculate_rand_source}
\usage{
calculate_rand_source(df_list, group)
}
\arguments{
\item{df_list}{[[list]] list of data.frames}

\item{group}{[[character]] group variable name \itemize{
\item option1
\item option2
\item option3}}
}
\value{
[[data.frame]]  summarised data.frame
}
\description{
Function calculates randomized subject number and ratio in groups
}
