% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_biggest_value}
\alias{extract_biggest_value}
\title{extract_biggest_value}
\usage{
extract_biggest_value(df, vars)
}
\arguments{
\item{df}{[[data.frame]] data}

\item{vars}{[[character]] variable names, ordered by the importance}
}
\value{
[[data.frame]]  1-row data.frame with the maximal value of the first chosen variable (and distant one in case of ties)
}
\description{
Function extracts the row with the maximal value
}
