% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fill_values_in_list}
\alias{fill_values_in_list}
\title{fill_values_in_list}
\usage{
fill_values_in_list(task_list, names_in, names_out, fun, groups = NULL, ...)
}
\arguments{
\item{task_list}{[[list]] data}

\item{names_in}{[[character]] list elements names used in calculations}

\item{names_out}{[[character]] list element names for calculated values}

\item{fun}{[[character]] function name, currently only `"extract_biggest_value"`}

\item{groups}{[[character]] vector with group variable names}

\item{...}{additional parameters}
}
\value{
[[list]] filled in/extended list
}
\description{
Function fills in or extend the given list
}
