% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{level2logical}
\alias{level2logical}
\title{level2logical}
\usage{
level2logical(df, var, level, negate = FALSE)
}
\arguments{
\item{df}{[[data.frame]] data}

\item{var}{[[character]] variable name to check the level presence}

\item{level}{[[character]]/[[numeric]] chosen level}
}
\value{
[[data.frame]] new data.frame with added variable
}
\description{
Function adds one logical variable indicating if the given level is observed
}
